+++ example/lib/utils/ui_helper.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class P5 extends StatelessWidget {
+class Padding5 extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(padding: EdgeInsets.all(5.0));