+++ lib/src/widgets/connectivity_app_wrapper_widget.dart
@@ -1,4 +1,5 @@
 import 'package:connectivity_wrapper/src/providers/connectivity_provider.dart';
+import 'package:connectivity_wrapper/src/utils/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -8,22 +9,14 @@ class ConnectivityAppWrapper extends StatelessWidget {
   /// [app] will accept MaterialApp or CupertinoApp must be non-null
   final Widget app;
 
-  /// [connectivityServiceRefreshTime] Defines the time interval
-  /// to check the ConnectivityStatus again
-  /// default[defaultRefreshTime] set to 2000 milliseconds
-  final int connectivityServiceRefreshTime;
-
-  const ConnectivityAppWrapper(
-      {Key key, @required this.app, this.connectivityServiceRefreshTime})
+  const ConnectivityAppWrapper({Key key, @required this.app})
       : assert(app != null),
         super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return StreamProvider<ConnectivityStatus>(
-      builder: (context) => ConnectivityProvider(
-              connectivityServiceRefreshTime: connectivityServiceRefreshTime)
-          .connectivityStream,
+      builder: (context) => ConnectivityProvider().connectivityStream,
       child: app,
     );
   }